+++ src/main/resources/CHANGELOG.md
@@ -1,3 +1,6 @@
+## Unreleased
+- Fixed failure to detect config version mismatch.
+
 ## 1.2.2
 - Refactored Version Checker integration to be reused by client mods.